# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Fantasy Yahoo Service API
  description: "Integrates with the Yahoo Fantasy Sports API: owns the per-user OAuth\
    \ 2.0 flow and exposes a user's fantasy league settings to the BFF."
  version: v1
servers:
- url: /
...
paths:
...
  /api/v1/yahoo/players:
    get:
      tags:
      - Yahoo Players
      summary: List every Yahoo player for a game with eligible positions and season
        stats
      description: "Paginates the Yahoo game player collection using the service account\
        \ token, returning each player's identity, headshot, eligible positions and\
        \ the requested season's stat line. Defaults to the NHL game; a blank season\
        \ means the current season."
      operationId: players
      parameters:
      - name: gameKey
        in: query
        required: false
        schema:
          type: string
          default: nhl
      - name: season
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Players returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YahooPlayerResponse"
        "404":
          description: The Yahoo service account is not connected
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YahooPlayerResponse"
...
  /api/v1/players/{playerId}/headshot:
    get:
      tags:
      - Players
      summary: A player's headshot thumbnail
      description: "The player's headshot, scaled to 192px tall and otherwise as Yahoo\
        \ frames it — the caller crops it to the avatar it draws. Refreshed by the\
        \ sync, so the response is safe to cache for a long time and is revalidated\
        \ with an ETag."
      operationId: headshot
      parameters:
      - name: playerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Thumbnail returned
          content:
            image/png:
              schema:
                type: string
                format: byte
        "404":
          description: No headshot stored for this player
          content:
            image/png:
              schema:
                type: string
                format: byte
  /api/v1/players/skaters:
    get:
      tags:
      - Players
      summary: Skaters with eligible positions and a season's stats
      description: "The whole pool, or its highest scoring when `limit` is given."
      operationId: skaters
      parameters:
      - name: season
        in: query
        description: "Season start year — 2025 is the 2025-26 season. Required: the\
          \ season being collected and the season a caller wants to show are deliberately\
          \ different for most of the year, so there is no sensible default. A player\
          \ with no line for it comes back with no stats rather than being left out."
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        description: "How many to return, highest scoring first (most wins, for goalies).\
          \ Absent returns the whole pool in name order, which is what a caller ranking\
          \ every player wants; a caller drawing a handful of rows should ask for\
          \ a handful."
        required: false
        schema:
          type: integer
          format: int32
          maximum: 500
          minimum: 1
      responses:
        "200":
          description: Skaters returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkaterResponse"
        "400":
          description: limit is not between 1 and 500
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkaterResponse"
  /api/v1/players/goalies:
    get:
      tags:
      - Players
      summary: Goalies with eligible positions and a season's stats
      description: "The whole pool, or its winningest when `limit` is given."
      operationId: goalies
      parameters:
      - name: season
        in: query
        description: "Season start year — 2025 is the 2025-26 season. Required: the\
          \ season being collected and the season a caller wants to show are deliberately\
          \ different for most of the year, so there is no sensible default. A player\
          \ with no line for it comes back with no stats rather than being left out."
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        description: "How many to return, highest scoring first (most wins, for goalies).\
          \ Absent returns the whole pool in name order, which is what a caller ranking\
          \ every player wants; a caller drawing a handful of rows should ask for\
          \ a handful."
        required: false
        schema:
          type: integer
          format: int32
          maximum: 500
          minimum: 1
      responses:
        "200":
          description: Goalies returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalieResponse"
        "400":
          description: limit is not between 1 and 500
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalieResponse"
components:
  schemas:
...
    YahooGoalieStats:
      type: object
      properties:
        gamesPlayed:
          type: integer
          format: int32
        gamesStarted:
          type: integer
          format: int32
        wins:
          type: integer
          format: int32
        losses:
          type: integer
          format: int32
        shutouts:
          type: integer
          format: int32
        shotsAgainst:
          type: integer
          format: int32
        saves:
          type: integer
          format: int32
        goalsAgainst:
          type: integer
          format: int32
        goalsAgainstAvg:
          type: number
          format: double
          description: Goals against average
        savePct:
          type: number
          format: double
          description: Save percentage as a fraction (e.g. 0.911)
    YahooPlayerResponse:
      type: object
      properties:
        yahooId:
          type: string
          description: Yahoo's player id
        fullName:
          type: string
          description: Player full name
        firstName:
          type: string
          description: Player first name
        lastName:
          type: string
          description: Player last name
        teamAbbrev:
          type: string
          description: Yahoo's editorial team abbreviation
        position:
          type: string
          description: "Primary position (single, e.g. \"C\")"
        uniformNumber:
          type: integer
          format: int32
          description: Uniform / sweater number
        imageUrl:
          type: string
          description: Full-resolution headshot image URL
        goalie:
          type: boolean
          description: Whether the player is a goalie
        eligiblePositions:
          type: array
          description: "Yahoo fantasy eligible positions (e.g. [\"C\",\"LW\"])"
          items:
            type: string
        skaterStats:
          $ref: "#/components/schemas/YahooSkaterStats"
          description: Season skater stat line; null for goalies
        goalieStats:
          $ref: "#/components/schemas/YahooGoalieStats"
          description: Season goalie stat line; null for skaters
      required:
      - eligiblePositions
      - firstName
      - fullName
      - goalie
      - lastName
      - yahooId
    YahooSkaterStats:
      type: object
      properties:
        gamesPlayed:
          type: integer
          format: int32
        goals:
          type: integer
          format: int32
        assists:
          type: integer
          format: int32
        points:
          type: integer
          format: int32
        plusMinus:
          type: integer
          format: int32
        pim:
          type: integer
          format: int32
        powerPlayGoals:
          type: integer
          format: int32
        powerPlayPoints:
          type: integer
          format: int32
        shorthandedGoals:
          type: integer
          format: int32
        shorthandedPoints:
          type: integer
          format: int32
        gameWinningGoals:
          type: integer
          format: int32
        shotsOnGoal:
          type: integer
          format: int32
        shootingPct:
          type: number
          format: double
          description: Shooting percentage as a fraction (e.g. 0.151)
        faceoffsWon:
          type: integer
          format: int32
        faceoffsLost:
          type: integer
          format: int32
        hits:
          type: integer
          format: int32
        blocks:
          type: integer
          format: int32
        avgTimeOnIce:
          type: string
          description: "Average time on ice per game, formatted \"MM:SS\""
...
    SkaterResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        eligiblePositions:
          type: array
          description: "Yahoo fantasy eligible positions (e.g. [\"C\",\"LW\"])"
          items:
            type: string
        sweaterNumber:
          type: integer
          format: int32
        teamAbbrev:
          type: string
        headshot:
          type: string
          description: "Yahoo's source image URL, present only when a thumbnail is\
            \ served at /api/v1/players/{id}/headshot — clients should render that\
            \ endpoint, not this URL"
        gamesPlayed:
          type: integer
          format: int32
        goals:
          type: integer
          format: int32
        assists:
          type: integer
          format: int32
        points:
          type: integer
          format: int32
        plusMinus:
          type: integer
          format: int32
        pim:
          type: integer
          format: int32
        powerPlayGoals:
          type: integer
          format: int32
        powerPlayPoints:
          type: integer
          format: int32
        shorthandedGoals:
          type: integer
          format: int32
        shorthandedPoints:
          type: integer
          format: int32
        gameWinningGoals:
          type: integer
          format: int32
        shots:
          type: integer
          format: int32
        shootingPctg:
          type: number
          format: double
        avgToi:
          type: string
        faceoffWinningPctg:
          type: number
          format: double
        hits:
          type: integer
          format: int32
        blockedShots:
          type: integer
          format: int32
        totalFaceoffWins:
          type: integer
          format: int32
        totalFaceoffLosses:
          type: integer
          format: int32
      required:
      - eligiblePositions
      - firstName
      - id
      - lastName
      - position
    GoalieResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        eligiblePositions:
          type: array
          description: "Yahoo fantasy eligible positions (e.g. [\"G\"])"
          items:
            type: string
        sweaterNumber:
          type: integer
          format: int32
        teamAbbrev:
          type: string
        headshot:
          type: string
          description: "Yahoo's source image URL, present only when a thumbnail is\
            \ served at /api/v1/players/{id}/headshot — clients should render that\
            \ endpoint, not this URL"
        gamesPlayed:
          type: integer
          format: int32
        gamesStarted:
          type: integer
          format: int32
        wins:
          type: integer
          format: int32
        losses:
          type: integer
          format: int32
        shutouts:
          type: integer
          format: int32
        shotsAgainst:
          type: integer
          format: int32
        saves:
          type: integer
          format: int32
        goalsAgainst:
          type: integer
          format: int32
        goalsAgainstAvg:
          type: number
          format: double
        savePctg:
          type: number
          format: double
      required:
      - eligiblePositions
      - firstName
      - id
      - lastName
      - position